+++ src/core/protocols/oauth/oauth2_callback.ts
@@ -1,9 +1,9 @@
 import type { PayloadRequest } from 'payload'
 import * as oauth from 'oauth4webapi'
-import type { OAuth2ProviderConfig, OAuthAccountInfo } from '../../types'
-import { getCallbackURL } from '../utils/cb'
-import { parseCookies } from '../utils/cookies'
-import { MissingOrInvalidSession } from '../error'
+import type { OAuth2ProviderConfig, OAuthAccountInfo } from '../../../types'
+import { getCallbackURL } from '../../utils/cb'
+import { parseCookies } from '../../utils/cookies'
+import { MissingOrInvalidSession } from '../../error'
 
 export async function OAuth2Callback(
   request: PayloadRequest,
