+++ src/core/routeHandlers/oauth.ts
@@ -1,10 +1,10 @@
 import type { PayloadRequest } from 'payload'
 import type { OAuthAccountInfo, OAuthProviderConfig } from '../../types'
 import { InvalidOAuthAlgorithm, InvalidOAuthResource, InvalidProvider } from '../error'
-import { OIDCAuthorization } from '../protocols/oidc_authorization'
-import { OAuth2Authorization } from '../protocols/oauth2_authorization'
-import { OIDCCallback } from '../protocols/oidc_callback'
-import { OAuth2Callback } from '../protocols/oauth2_callback'
+import { OIDCAuthorization } from '../protocols/oauth/oidc_authorization'
+import { OAuth2Authorization } from '../protocols/oauth/oauth2_authorization'
+import { OIDCCallback } from '../protocols/oauth/oidc_callback'
+import { OAuth2Callback } from '../protocols/oauth/oauth2_callback'
 
 export function OAuthHandlers(
   request: PayloadRequest,
